requirements.txt
dnspython==2.3.0
paho-mqtt==1.6.1
pymongo==4.3.3
